src/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.11.0
python-multipart==0.0.20
numpy==2.2.6
pandas==2.3.2
joblib==1.5.2
scikit-learn==1.7.1
torch==2.8.0
torchvision==0.23.0
Pillow==11.3.0
git+https://github.com/openai/CLIP.git
